refactor(profile): shared endorsement subject row, single hook call, stable callbacks; split below-fold + modals
EndorsementSubjectRow (on deriveIdentity, info+isLoading props) replaces
three drifted row copies; ProfileEndorsements calls useGivenEndorsements
once and derives the endorsed-set from it; row onToggle handlers are
useCallback-stable so list memos hold; below-fold views split to
profile-endorsement-views.tsx and the three list modals to their own
module.
