+++ src/components/endorsements/endorsement-row.tsx
@@ -1,15 +1,13 @@
 "use client"
 
-import Link from "next/link"
-import { profileUrl } from "@/lib/urls"
 import { X } from "lucide-react"
-import Avatar from "@/components/ui/avatar"
 import LoadingSpinner from "@/components/ui/loading-spinner"
-import Skeleton from "@/components/ui/skeleton"
 import Tooltip from "@/components/ui/tooltip"
+import EndorsementSubjectRow, {
+  type EndorsementSubjectRowClasses,
+} from "@/components/endorsements/endorsement-subject-row"
 import { useAuthorInfo } from "@/hooks/use-author-info"
-import { getInitials } from "@/lib/utils/initials"
-import { formatShortDate } from "@/lib/utils/format-date"
+import { deriveIdentity } from "@/lib/utils/identity"
 
 interface EndorsementRowProps {
   /** The DID of the endorsed account. */
@@ -27,11 +25,21 @@ interface EndorsementRowProps {
   readonly isRevoking?: boolean
 }
 
+const ROW_CLASSES: EndorsementSubjectRowClasses = {
+  main: "endorsement-row__main",
+  meta: "endorsement-row__meta",
+  name: "endorsement-row__name",
+  handle: "endorsement-row__handle",
+  note: "endorsement-row__note",
+  date: "endorsement-row__date",
+}
+
 /**
  * Single row in an endorsements list. Hydrates the subject DID into
  * avatar + display name + handle via `useAuthorInfo` (same hook
- * powering activity card bylines), and links through to the
- * subject's profile page. Optionally shows a revoke button.
+ * powering activity card bylines) and renders the shared
+ * `EndorsementSubjectRow`, which links through to the subject's
+ * profile page. Optionally shows a revoke button.
  */
 export default function EndorsementRow({
   subjectDid,
@@ -41,53 +49,35 @@ export default function EndorsementRow({
   isRevoking,
 }: EndorsementRowProps) {
   const { info, isLoading } = useAuthorInfo(subjectDid)
-
-  const displayName = info?.displayName || info?.handle || subjectDid
-  const handle = info?.handle && info.handle !== info.did ? info.handle : null
-  const initials = getInitials(info?.displayName, subjectDid)
-  const href = profileUrl(info?.handle || subjectDid)
+  // Same derivation the shared row renders, so the revoke aria-label
+  // matches the visible display name.
+  const { displayName } = deriveIdentity(info, subjectDid)
 
   return (
     <li className="endorsement-row">
-      <Link href={href} className="endorsement-row__main">
-        {isLoading && !info ? (
-          <Skeleton circle animate={false} width={48} height={48} />
-        ) : (
-          <Avatar
-            size="md"
-            src={info?.avatarUrl || undefined}
-            alt=""
-            fallbackInitials={initials}
-          />
-        )}
-        <div className="endorsement-row__meta">
-          <span className="endorsement-row__name">{displayName}</span>
-          {handle ? (
-            <span className="endorsement-row__handle">@{handle}</span>
-          ) : null}
-          {note ? <span className="endorsement-row__note">{note}</span> : null}
-        </div>
-      </Link>
-      <time
-        dateTime={createdAt}
-        className="endorsement-row__date"
-        title={new Date(createdAt).toLocaleString()}
-      >
-        {formatShortDate(createdAt)}
-      </time>
-      {onRevoke ? (
-        <Tooltip label="Revoke endorsement">
-          <button
-            type="button"
-            className="endorsement-row__revoke"
-            onClick={onRevoke}
-            disabled={isRevoking}
-            aria-label={`Revoke endorsement of ${displayName}`}
-          >
-            {isRevoking ? <LoadingSpinner size="sm" /> : <X size={16} />}
-          </button>
-        </Tooltip>
-      ) : null}
+      <EndorsementSubjectRow
+        did={subjectDid}
+        info={info}
+        isLoading={isLoading}
+        createdAt={createdAt}
+        note={note}
+        classes={ROW_CLASSES}
+        trailing={
+          onRevoke ? (
+            <Tooltip label="Revoke endorsement">
+              <button
+                type="button"
+                className="endorsement-row__revoke"
+                onClick={onRevoke}
+                disabled={isRevoking}
+                aria-label={`Revoke endorsement of ${displayName}`}
+              >
+                {isRevoking ? <LoadingSpinner size="sm" /> : <X size={16} />}
+              </button>
+            </Tooltip>
+          ) : null
+        }
+      />
     </li>
   )
 }
