+++ src/components/profile/profile-endorsements.tsx
@@ -1,30 +1,21 @@
 "use client"
 
-import { memo, useCallback, useEffect, useMemo, useState } from "react"
+import { useCallback, useEffect, useMemo, useState } from "react"
 import { useUrlParam } from "@/hooks/use-url-param"
 import {
   ArrowUpDown,
   Ban,
   Filter,
-  Inbox,
   LayoutGrid,
   List as ListIcon,
   Plus,
   Search,
-  ThumbsUp,
   Trash2,
-  X,
-  type LucideIcon,
 } from "lucide-react"
-import { useGivenEndorsements, type GivenEndorsement } from "@/hooks/use-endorsements"
-import {
-  useReceivedEndorsements,
-  type ReceivedEndorsement,
-} from "@/hooks/use-received-endorsements"
+import { useGivenEndorsements } from "@/hooks/use-endorsements"
+import { useReceivedEndorsements } from "@/hooks/use-received-endorsements"
 import { useOwnResponseStates } from "@/hooks/use-own-response-states"
-import { useAuthorInfo, type AuthorInfo } from "@/hooks/use-author-info"
 import { useAuthorNamesMap } from "@/hooks/use-author-names-map"
-import { buildAvatarUrlFromCid } from "@/lib/atproto/profile"
 import { useAuth } from "@/lib/auth/auth-context"
 import { useOrg } from "@/lib/groups/org-context"
 import {
@@ -32,20 +23,21 @@ import {
   deleteEndorsementAward,
   createResponse,
 } from "@/lib/atproto/badges"
-import ResponseMenu from "@/components/badges/response-menu"
-import Avatar from "@/components/ui/avatar"
-import Button from "@/components/ui/button"
-import Checkbox from "@/components/ui/checkbox"
 import SegmentedControl from "@/components/ui/segmented-control"
-import { profileUrl } from "@/lib/urls"
-import { getInitials } from "@/lib/utils/initials"
-import Link from "next/link"
 import EndorsementLists from "@/components/profile/endorsement-lists"
 import EndorsePeopleModal from "@/components/profile/endorse-people-modal"
-import PersonCard from "@/components/profile/person-card"
+import {
+  BulkBar,
+  GivenGrid,
+  GivenList,
+  ReceivedGrid,
+  ReceivedList,
+  filterAndSortGiven,
+  filterAndSortReceived,
+  type ResponseFilterKey,
+  type SortKey,
+} from "@/components/profile/profile-endorsement-views"
 import ConfirmDialog from "@/components/ui/confirm-dialog"
-import EmptyState from "@/components/ui/empty-state"
-import LoadingSpinner from "@/components/ui/loading-spinner"
 import Badge from "@/components/ui/badge"
 import {
   Popover,
@@ -64,12 +56,6 @@ interface ProfileEndorsementsProps {
 type SubTab = "received" | "given"
 type ViewMode = "gallery" | "list"
 
-type SortKey =
-  | "created-desc"
-  | "created-asc"
-  | "alpha-asc"
-  | "alpha-desc"
-
 /** Sort options shown in the toolbar. Labels swap between sub-tabs
  *  because A→Z means different things ("Endorser" vs "Recipient"). */
 const RECEIVED_SORT_OPTIONS: { key: SortKey; label: string }[] = [
@@ -86,7 +72,6 @@ const GIVEN_SORT_OPTIONS: { key: SortKey; label: string }[] = [
   { key: "alpha-desc", label: "Recipient Z → A" },
 ]
 
-type ResponseFilterKey = "hide-rejected" | "only-rejected" | "show-all"
 const RESPONSE_FILTER_OPTIONS: { key: ResponseFilterKey; label: string }[] = [
   { key: "hide-rejected", label: "Hide rejected" },
   { key: "only-rejected", label: "Show only rejected" },
@@ -211,14 +196,15 @@ export default function ProfileEndorsements({ did }: ProfileEndorsementsProps) {
 
   // Endorse-people modal (own-profile only). The viewer can search
   // for one or many people and write a batch of endorsements in a
-  // single pass.
+  // single pass. When the viewer can manage this profile, `given` IS
+  // their own endorsement set — no second useGivenEndorsements call
+  // (the hook has no cache; each instance re-runs the whole
+  // listDefinitions/listAwards/per-definition fan-out).
   const [isEndorseModalOpen, setIsEndorseModalOpen] = useState(false)
-  const ownGivenForModal = useGivenEndorsements(
-    canManage ? did : null,
-  )
   const ownAlreadyEndorsedDids = useMemo(
-    () => new Set(ownGivenForModal.endorsements.map((e) => e.subjectDid)),
-    [ownGivenForModal.endorsements],
+    () =>
+      new Set(canManage ? given.endorsements.map((e) => e.subjectDid) : []),
+    [canManage, given.endorsements],
   )
 
   // The DID set to hydrate names for — issuers for the Received tab,
@@ -632,7 +618,6 @@ export default function ProfileEndorsements({ did }: ProfileEndorsementsProps) {
             // its own endorsements — no refresh needed here. Close
             // after the refresh so the user sees the spinner clear.
             setIsEndorseModalOpen(false)
-            await ownGivenForModal.refetch()
             await given.refetch()
           }}
         />
@@ -644,834 +629,4 @@ export default function ProfileEndorsements({ did }: ProfileEndorsementsProps) {
 function formatCount(n: number): string | null {
   if (n === 0) return null
   return `${n}`
-}
-
-// ----------------------------- Received -----------------------------
-
-interface ReceivedGridProps {
-  /** Already filtered + sorted by the parent (`visibleReceived`); the
-   *  grid renders it directly instead of recomputing the sort. */
-  visible: ReceivedEndorsement[]
-  /** Size of the pre-search set — drives the "No endorsements yet" vs
-   *  "No matches" empty-state split. */
-  total: number
-  isLoading: boolean
-  error: string | null
-  viewerIsOwner: boolean
-  viewerDid: string | null
-  /** Group DID when acting AS this group — accept/reject responses route
-   *  to the group's repo. Undefined for personal responses. */
-  targetDid?: string
-  /** Active response filter — used by the empty-state copy so a
-   *  zero-results "Show only rejected" view says "No rejected
-   *  endorsements yet" instead of the generic "No endorsements
-   *  yet." */
-  responseFilter: ResponseFilterKey
-  resolve: ReturnType<typeof useOwnResponseStates>["resolve"]
-  allResponses: ReturnType<typeof useOwnResponseStates>["responses"]
-  onAfterWrite: () => void | Promise<void>
-}
-
-function ReceivedGrid({
-  visible,
-  total,
-  isLoading,
-  error,
-  viewerIsOwner,
-  viewerDid,
-  targetDid,
-  responseFilter,
-  resolve,
-  allResponses,
-  onAfterWrite,
-}: ReceivedGridProps) {
-  if (isLoading) {
-    return (
-      <div className="profile-endorsements-v2__loading">
-        <LoadingSpinner size="md" />
-      </div>
-    )
-  }
-  if (error) {
-    return (
-      <div className="profile-endorsements-v2__grid">
-        <EmptyState
-          icon={Inbox}
-          title="Couldn’t load endorsements"
-          description={error}
-        />
-      </div>
-    )
-  }
-  if (visible.length === 0) {
-    // Three empty-state cases:
-    //   1. The "Show only rejected" filter is active with no matches
-    //      — phrase the empty state in terms of the filter so the
-    //      user knows nothing is missing, the filter just has no
-    //      hits yet.
-    //   2. There's a search query / sort filter but the pre-filter
-    //      set is non-empty — "No matches."
-    //   3. The user has zero endorsements total — the generic
-    //      "No endorsements yet" CTA.
-    const onlyRejectedActive = responseFilter === "only-rejected"
-    const title = onlyRejectedActive
-      ? "No rejected endorsements yet"
-      : total === 0
-        ? "No endorsements yet"
-        : "No matches"
-    const description = onlyRejectedActive
-      ? "Endorsements you reject will appear here."
-      : total === 0
-        ? "Endorsements from other people will appear here."
-        : "No endorsements match your search."
-    return (
-      <div className="profile-endorsements-v2__grid">
-        <EmptyState icon={ThumbsUp} title={title} description={description} />
-      </div>
-    )
-  }
-  return (
-    <ul className="profile-endorsements-v2__grid">
-      {visible.map((e) => (
-        <ReceivedCard
-          key={e.uri}
-          endorsement={e}
-          viewerIsOwner={viewerIsOwner}
-          viewerDid={viewerDid}
-          targetDid={targetDid}
-          resolve={resolve}
-          allResponses={allResponses}
-          onAfterWrite={onAfterWrite}
-        />
-      ))}
-    </ul>
-  )
-}
-
-const ReceivedCard = memo(function ReceivedCard({
-  endorsement,
-  viewerIsOwner,
-  viewerDid,
-  targetDid,
-  resolve,
-  allResponses,
-  onAfterWrite,
-}: {
-  endorsement: ReceivedEndorsement
-  viewerIsOwner: boolean
-  viewerDid: string | null
-  targetDid?: string
-  resolve: ReturnType<typeof useOwnResponseStates>["resolve"]
-  allResponses: ReturnType<typeof useOwnResponseStates>["responses"]
-  onAfterWrite: () => void | Promise<void>
-}) {
-  const { info, isLoading } = useReceivedIssuerInfo(endorsement)
-
-  return (
-    <PersonCard
-      did={endorsement.issuerDid}
-      info={info}
-      isLoadingInfo={isLoading}
-      createdAt={endorsement.createdAt}
-      note={endorsement.note}
-      listTitle={endorsement.listTitle}
-      menu={
-        viewerIsOwner ? (
-          <ResponseMenu
-            awardUri={endorsement.uri}
-            awardCid={endorsement.cid}
-            issuerDisplayName={
-              info?.displayName || info?.handle || endorsement.issuerDid
-            }
-            ownerDid={viewerDid}
-            targetDid={targetDid}
-            state={resolve(endorsement.uri).state}
-            allResponses={allResponses}
-            onAfterWrite={onAfterWrite}
-          />
-        ) : null
-      }
-    />
-  )
-})
-
-// ------------------------------ Given -------------------------------
-
-interface GivenGridProps {
-  /** Already filtered + sorted by the parent (`visibleGiven`); the grid
-   *  renders it directly instead of recomputing the sort. */
-  visible: GivenEndorsement[]
-  /** Size of the pre-search set — drives the "No endorsements given yet"
-   *  vs "No matches" empty-state split. */
-  total: number
-  isLoading: boolean
-  error: string | null
-  /** True when the profile being viewed is the signed-in user's
-   *  own profile — i.e. the cards represent endorsements THEY
-   *  issued. Controls whether the per-card revoke `×` renders. */
-  viewerIsOwner: boolean
-  viewerDid: string | null
-  /** Group DID when the viewer is acting AS this group — revokes route
-   *  to the group repo. Undefined for personal revokes. */
-  targetDid?: string
-  onAfterRevoke: () => void | Promise<void>
-}
-
-function GivenGrid({
-  visible,
-  total,
-  isLoading,
-  error,
-  viewerIsOwner,
-  viewerDid,
-  targetDid,
-  onAfterRevoke,
-}: GivenGridProps) {
-  if (isLoading) {
-    return (
-      <div className="profile-endorsements-v2__loading">
-        <LoadingSpinner size="md" />
-      </div>
-    )
-  }
-  if (error) {
-    return (
-      <div className="profile-endorsements-v2__grid">
-        <EmptyState
-          icon={Inbox}
-          title="Couldn’t load endorsements"
-          description={error}
-        />
-      </div>
-    )
-  }
-  if (visible.length === 0) {
-    return (
-      <div className="profile-endorsements-v2__grid">
-        <EmptyState
-          icon={ThumbsUp}
-          title={total === 0 ? "No endorsements given yet" : "No matches"}
-          description={
-            total === 0
-              ? "Endorsements this user gives to others will appear here."
-              : "No endorsements match your search."
-          }
-        />
-      </div>
-    )
-  }
-  return (
-    <ul className="profile-endorsements-v2__grid">
-      {visible.map((e) => (
-        <GivenCard
-          key={e.uri}
-          endorsement={e}
-          canRevoke={viewerIsOwner && !!viewerDid}
-          viewerDid={viewerDid}
-          targetDid={targetDid}
-          onAfterRevoke={onAfterRevoke}
-        />
-      ))}
-    </ul>
-  )
-}
-
-const GivenCard = memo(function GivenCard({
-  endorsement,
-  canRevoke,
-  viewerDid,
-  targetDid,
-  onAfterRevoke,
-}: {
-  endorsement: GivenEndorsement
-  canRevoke: boolean
-  viewerDid: string | null
-  targetDid?: string
-  onAfterRevoke: () => void | Promise<void>
-}) {
-  const { info, isLoading } = useAuthorInfo(endorsement.subjectDid)
-  return (
-    <PersonCard
-      did={endorsement.subjectDid}
-      info={info}
-      isLoadingInfo={isLoading}
-      createdAt={endorsement.createdAt}
-      note={endorsement.note}
-      listTitle={endorsement.listTitle}
-      menu={
-        canRevoke && viewerDid ? (
-          <RevokeGivenButton
-            viewerDid={viewerDid}
-            rkeys={endorsement.rkeys}
-            targetDid={targetDid}
-            subjectDisplay={
-              info?.displayName || info?.handle || endorsement.subjectDid
-            }
-            onAfterRevoke={onAfterRevoke}
-          />
-        ) : null
-      }
-    />
-  )
-})
-
-/**
- * Small `×` revoke affordance shown on the owner's Given grid.
- * Click → ConfirmDialog ("Revoke endorsement?") → on confirm,
- * `deleteEndorsementAward` runs and the parent's `onAfterRevoke`
- * refetches the Given list so the card disappears.
- */
-function RevokeGivenButton({
-  viewerDid,
-  rkeys,
-  targetDid,
-  subjectDisplay,
-  onAfterRevoke,
-}: {
-  viewerDid: string
-  /** All award rkeys for this recipient — revoke removes every one so a
-   *  recipient endorsed more than once disappears in a single click. */
-  rkeys: string[]
-  targetDid?: string
-  subjectDisplay: string
-  onAfterRevoke: () => void | Promise<void>
-}) {
-  const [confirmOpen, setConfirmOpen] = useState(false)
-  const [isRevoking, setIsRevoking] = useState(false)
-  const [error, setError] = useState<string | null>(null)
-
-  const handleConfirm = async () => {
-    if (isRevoking) return
-    setIsRevoking(true)
-    setError(null)
-    try {
-      for (const rkey of rkeys) {
-        await deleteEndorsementAward(viewerDid, rkey, { targetDid })
-      }
-      await onAfterRevoke()
-      setConfirmOpen(false)
-    } catch (err) {
-      setError(err instanceof Error ? err.message : "Failed to revoke")
-    } finally {
-      setIsRevoking(false)
-    }
-  }
-
-  return (
-    <>
-      <Tooltip label="Revoke endorsement">
-        <button
-          type="button"
-          className="profile-endorsements-v2__given-revoke"
-          onClick={(e) => {
-            // PersonCard's outer Link otherwise catches the click and
-            // navigates to the subject's profile.
-            e.preventDefault()
-            e.stopPropagation()
-            setConfirmOpen(true)
-          }}
-          aria-label={`Revoke endorsement of ${subjectDisplay}`}
-        >
-          <X size={14} strokeWidth={2} aria-hidden />
-        </button>
-      </Tooltip>
-      {confirmOpen ? (
-        <ConfirmDialog
-          title={`Revoke endorsement of ${subjectDisplay}?`}
-          message="Your endorsement will be removed from this profile. You can endorse them again later."
-          confirmLabel="Revoke"
-          cancelLabel="Keep endorsement"
-          confirmVariant="destructive"
-          isConfirming={isRevoking}
-          onConfirm={handleConfirm}
-          onCancel={() => !isRevoking && setConfirmOpen(false)}
-        />
-      ) : null}
-      {error ? (
-        <span className="profile-endorsements-v2__endorse-error" role="alert">
-          {error}
-        </span>
-      ) : null}
-    </>
-  )
-}
-
-// ----------------------- List view + bulk select -----------------------
-
-/**
- * Compose the issuer's AuthorInfo from the indexer's denormalised block,
- * filling gaps via `useAuthorInfo`. The indexer's `issuer` join can carry
- * a handle WITHOUT a displayName/avatar (certified-only orgs have no
- * bsky-profile join), so we only skip the per-row resolve when the
- * indexer block is complete. Shared by the received card + list row.
- */
-function useReceivedIssuerInfo(endorsement: ReceivedEndorsement): {
-  info: AuthorInfo | null
-  isLoading: boolean
-} {
-  const idxIssuer = endorsement.issuer
-  const indexerAvatar = buildAvatarUrlFromCid(
-    idxIssuer?.did ?? endorsement.issuerDid,
-    idxIssuer?.avatarCid,
-  )
-  const indexerComplete = !!(
-    idxIssuer?.handle &&
-    idxIssuer.displayName &&
-    indexerAvatar
-  )
-  const { info: fetched, isLoading } = useAuthorInfo(
-    indexerComplete ? null : endorsement.issuerDid,
-  )
-  const info: AuthorInfo | null =
-    idxIssuer?.handle || idxIssuer?.displayName || indexerAvatar || fetched
-      ? {
-          did: idxIssuer?.did ?? endorsement.issuerDid,
-          handle: idxIssuer?.handle ?? fetched?.handle ?? endorsement.issuerDid,
-          displayName: idxIssuer?.displayName ?? fetched?.displayName ?? null,
-          avatarUrl: indexerAvatar ?? fetched?.avatarUrl ?? null,
-        }
-      : fetched
-  return { info, isLoading }
-}
-
-/** Select-all + bulk-action strip shown above the list view for owners. */
-function BulkBar({
-  selectedCount,
-  allSelected,
-  anyVisible,
-  onToggleAll,
-  actionLabel,
-  actionIcon: ActionIcon,
-  busy,
-  error,
-  onAction,
-}: {
-  selectedCount: number
-  allSelected: boolean
-  anyVisible: boolean
-  onToggleAll: () => void
-  actionLabel: string
-  actionIcon: LucideIcon
-  busy: boolean
-  error: string | null
-  onAction: () => void
-}) {
-  return (
-    <div
-      className="profile-endorsements-v2__bulk-bar"
-      role="toolbar"
-      aria-label="Bulk actions"
-    >
-      <Checkbox
-        checked={allSelected}
-        indeterminate={!allSelected && selectedCount > 0}
-        onChange={onToggleAll}
-        disabled={!anyVisible}
-        aria-label={allSelected ? "Deselect all" : "Select all"}
-        label={selectedCount > 0 ? `${selectedCount} selected` : "Select all"}
-      />
-      <div className="profile-endorsements-v2__bulk-actions">
-        {error ? (
-          <span className="profile-endorsements-v2__bulk-error" role="alert">
-            {error}
-          </span>
-        ) : null}
-        <Button
-          variant="destructive"
-          size="sm"
-          onClick={onAction}
-          disabled={selectedCount === 0 || busy}
-        >
-          <ActionIcon size={14} strokeWidth={1.75} aria-hidden />
-          {actionLabel}
-          {selectedCount > 0 ? ` (${selectedCount})` : ""}
-        </Button>
-      </div>
-    </div>
-  )
-}
-
-/** Shared list-row body: avatar + name/handle/note linking to the profile,
- *  with the date right-aligned. Used by both Given and Received rows. */
-function EndorsementRowBody({
-  did,
-  info,
-  createdAt,
-  note,
-}: {
-  did: string
-  info: AuthorInfo | null
-  createdAt: string
-  note?: string
-}) {
-  const display = info?.displayName || info?.handle || did
-  const handle =
-    info?.handle && info.handle !== info.did ? `@${info.handle}` : null
-  return (
-    <Link
-      href={profileUrl(info?.handle || did)}
-      className="profile-endorsements-v2__row-main"
-    >
-      <Avatar
-        size="sm"
-        src={info?.avatarUrl ?? undefined}
-        alt=""
-        fallbackInitials={getInitials(info?.displayName, info?.handle ?? did)}
-      />
-      <span className="profile-endorsements-v2__row-text">
-        <span className="profile-endorsements-v2__row-name">{display}</span>
-        {handle ? (
-          <span className="profile-endorsements-v2__row-handle">{handle}</span>
-        ) : null}
-        {note ? (
-          <span className="profile-endorsements-v2__row-note">{note}</span>
-        ) : null}
-      </span>
-      <time className="profile-endorsements-v2__row-date">
-        {createdAt.slice(0, 10)}
-      </time>
-    </Link>
-  )
-}
-
-function GivenList({
-  visible,
-  total,
-  isLoading,
-  error,
-  viewerIsOwner,
-  viewerDid,
-  targetDid,
-  selectable,
-  selected,
-  onToggleOne,
-  onAfterRevoke,
-}: {
-  visible: GivenEndorsement[]
-  total: number
-  isLoading: boolean
-  error: string | null
-  viewerIsOwner: boolean
-  viewerDid: string | null
-  targetDid?: string
-  selectable: boolean
-  selected: Set<string>
-  onToggleOne: (uri: string) => void
-  onAfterRevoke: () => void | Promise<void>
-}) {
-  if (isLoading) {
-    return (
-      <div className="profile-endorsements-v2__loading">
-        <LoadingSpinner size="md" />
-      </div>
-    )
-  }
-  if (error) {
-    return (
-      <EmptyState icon={Inbox} title="Couldn’t load endorsements" description={error} />
-    )
-  }
-  if (visible.length === 0) {
-    return (
-      <EmptyState
-        icon={ThumbsUp}
-        title={total === 0 ? "No endorsements given yet" : "No matches"}
-        description={
-          total === 0
-            ? "Endorsements this user gives to others will appear here."
-            : "No endorsements match your search."
-        }
-      />
-    )
-  }
-  return (
-    <ul className="profile-endorsements-v2__list">
-      {visible.map((e) => (
-        <GivenListRow
-          key={e.uri}
-          endorsement={e}
-          selectable={selectable}
-          selected={selected.has(e.uri)}
-          onToggle={() => onToggleOne(e.uri)}
-          canRevoke={viewerIsOwner && !!viewerDid}
-          viewerDid={viewerDid}
-          targetDid={targetDid}
-          onAfterRevoke={onAfterRevoke}
-        />
-      ))}
-    </ul>
-  )
-}
-
-const GivenListRow = memo(function GivenListRow({
-  endorsement,
-  selectable,
-  selected,
-  onToggle,
-  canRevoke,
-  viewerDid,
-  targetDid,
-  onAfterRevoke,
-}: {
-  endorsement: GivenEndorsement
-  selectable: boolean
-  selected: boolean
-  onToggle: () => void
-  canRevoke: boolean
-  viewerDid: string | null
-  targetDid?: string
-  onAfterRevoke: () => void | Promise<void>
-}) {
-  const { info } = useAuthorInfo(endorsement.subjectDid)
-  const display = info?.displayName || info?.handle || endorsement.subjectDid
-  return (
-    <li className="profile-endorsements-v2__row" data-selected={selected || undefined}>
-      {selectable ? (
-        <Checkbox
-          className="profile-endorsements-v2__row-check"
-          checked={selected}
-          onChange={onToggle}
-          aria-label={`Select endorsement of ${display}`}
-        />
-      ) : null}
-      <EndorsementRowBody
-        did={endorsement.subjectDid}
-        info={info}
-        createdAt={endorsement.createdAt}
-        note={endorsement.note}
-      />
-      {canRevoke && viewerDid ? (
-        <RevokeGivenButton
-          viewerDid={viewerDid}
-          rkeys={endorsement.rkeys}
-          targetDid={targetDid}
-          subjectDisplay={display}
-          onAfterRevoke={onAfterRevoke}
-        />
-      ) : null}
-    </li>
-  )
-})
-
-function ReceivedList({
-  visible,
-  total,
-  isLoading,
-  error,
-  responseFilter,
-  viewerIsOwner,
-  viewerDid,
-  targetDid,
-  resolve,
-  allResponses,
-  selectable,
-  selected,
-  onToggleOne,
-  onAfterWrite,
-}: {
-  visible: ReceivedEndorsement[]
-  total: number
-  isLoading: boolean
-  error: string | null
-  responseFilter: ResponseFilterKey
-  viewerIsOwner: boolean
-  viewerDid: string | null
-  targetDid?: string
-  resolve: ReturnType<typeof useOwnResponseStates>["resolve"]
-  allResponses: ReturnType<typeof useOwnResponseStates>["responses"]
-  selectable: boolean
-  selected: Set<string>
-  onToggleOne: (uri: string) => void
-  onAfterWrite: () => void | Promise<void>
-}) {
-  if (isLoading) {
-    return (
-      <div className="profile-endorsements-v2__loading">
-        <LoadingSpinner size="md" />
-      </div>
-    )
-  }
-  if (error) {
-    return (
-      <EmptyState icon={Inbox} title="Couldn’t load endorsements" description={error} />
-    )
-  }
-  if (visible.length === 0) {
-    const onlyRejectedActive = responseFilter === "only-rejected"
-    return (
-      <EmptyState
-        icon={ThumbsUp}
-        title={
-          onlyRejectedActive
-            ? "No rejected endorsements yet"
-            : total === 0
-              ? "No endorsements yet"
-              : "No matches"
-        }
-        description={
-          onlyRejectedActive
-            ? "Endorsements you reject will appear here."
-            : total === 0
-              ? "Endorsements from other people will appear here."
-              : "No endorsements match your search."
-        }
-      />
-    )
-  }
-  return (
-    <ul className="profile-endorsements-v2__list">
-      {visible.map((e) => (
-        <ReceivedListRow
-          key={e.uri}
-          endorsement={e}
-          selectable={selectable}
-          selected={selected.has(e.uri)}
-          onToggle={() => onToggleOne(e.uri)}
-          viewerIsOwner={viewerIsOwner}
-          viewerDid={viewerDid}
-          targetDid={targetDid}
-          resolve={resolve}
-          allResponses={allResponses}
-          onAfterWrite={onAfterWrite}
-        />
-      ))}
-    </ul>
-  )
-}
-
-const ReceivedListRow = memo(function ReceivedListRow({
-  endorsement,
-  selectable,
-  selected,
-  onToggle,
-  viewerIsOwner,
-  viewerDid,
-  targetDid,
-  resolve,
-  allResponses,
-  onAfterWrite,
-}: {
-  endorsement: ReceivedEndorsement
-  selectable: boolean
-  selected: boolean
-  onToggle: () => void
-  viewerIsOwner: boolean
-  viewerDid: string | null
-  targetDid?: string
-  resolve: ReturnType<typeof useOwnResponseStates>["resolve"]
-  allResponses: ReturnType<typeof useOwnResponseStates>["responses"]
-  onAfterWrite: () => void | Promise<void>
-}) {
-  const { info } = useReceivedIssuerInfo(endorsement)
-  const display = info?.displayName || info?.handle || endorsement.issuerDid
-  return (
-    <li
-      className="profile-endorsements-v2__row"
-      data-selected={selected || undefined}
-      data-state={resolve(endorsement.uri).state}
-    >
-      {selectable ? (
-        <Checkbox
-          className="profile-endorsements-v2__row-check"
-          checked={selected}
-          onChange={onToggle}
-          aria-label={`Select endorsement from ${display}`}
-        />
-      ) : null}
-      <EndorsementRowBody
-        did={endorsement.issuerDid}
-        info={info}
-        createdAt={endorsement.createdAt}
-        note={endorsement.note}
-      />
-      {viewerIsOwner ? (
-        <ResponseMenu
-          awardUri={endorsement.uri}
-          awardCid={endorsement.cid}
-          issuerDisplayName={display}
-          ownerDid={viewerDid}
-          targetDid={targetDid}
-          state={resolve(endorsement.uri).state}
-          allResponses={allResponses}
-          onAfterWrite={onAfterWrite}
-        />
-      ) : null}
-    </li>
-  )
-})
-
-// ----------------------- Filter + sort helpers -----------------------
-
-function filterAndSortReceived(
-  records: ReceivedEndorsement[],
-  query: string,
-  sort: SortKey,
-  names: Map<string, string>,
-): ReceivedEndorsement[] {
-  const q = query.trim().toLowerCase()
-  const matches = q
-    ? records.filter((r) => {
-        const note = (r.note ?? "").toLowerCase()
-        const name = names.get(r.issuerDid) ?? r.issuerDid.toLowerCase()
-        return note.includes(q) || name.includes(q)
-      })
-    : records
-
-  const sorted = matches.slice()
-  sorted.sort((a, b) => {
-    switch (sort) {
-      case "created-desc":
-        return compareString(b.createdAt, a.createdAt)
-      case "created-asc":
-        return compareString(a.createdAt, b.createdAt)
-      case "alpha-asc":
-        return (names.get(a.issuerDid) ?? "").localeCompare(
-          names.get(b.issuerDid) ?? "",
-        )
-      case "alpha-desc":
-        return (names.get(b.issuerDid) ?? "").localeCompare(
-          names.get(a.issuerDid) ?? "",
-        )
-    }
-  })
-  return sorted
-}
-
-function filterAndSortGiven(
-  records: GivenEndorsement[],
-  query: string,
-  sort: SortKey,
-  names: Map<string, string>,
-): GivenEndorsement[] {
-  const q = query.trim().toLowerCase()
-  const matches = q
-    ? records.filter((r) => {
-        const note = (r.note ?? "").toLowerCase()
-        const name = names.get(r.subjectDid) ?? r.subjectDid.toLowerCase()
-        return note.includes(q) || name.includes(q)
-      })
-    : records
-
-  const sorted = matches.slice()
-  sorted.sort((a, b) => {
-    switch (sort) {
-      case "created-desc":
-        return compareString(b.createdAt, a.createdAt)
-      case "created-asc":
-        return compareString(a.createdAt, b.createdAt)
-      case "alpha-asc":
-        return (names.get(a.subjectDid) ?? "").localeCompare(
-          names.get(b.subjectDid) ?? "",
-        )
-      case "alpha-desc":
-        return (names.get(b.subjectDid) ?? "").localeCompare(
-          names.get(a.subjectDid) ?? "",
-        )
-    }
-  })
-  return sorted
-}
-
-function compareString(a: string, b: string): number {
-  return a < b ? -1 : a > b ? 1 : 0
 }
